Implementación de la capa de servicio necesaria hasta el momento; refactorización de tipos de datos en gateways y deserializadores. Implementación al 70% del cron y del parseado de los datos. Actualmente queda implementar el servicio de email, parsear los precios y la persistencia en base de datos.

+++ src/main/java/com/inggarciabaldo/carburo/application/persistance/ccaa/CCAAGateway.java
@@ -6,8 +6,8 @@ import com.inggarciabaldo.carburo.application.persistance.ccaa.CCAAGateway.CCAAR
 public interface CCAAGateway extends Gateway<CCAARecord> {
 
 	class CCAARecord {
-		public int id;
+		public short id;
 		public String denominacion;
-		public int extCode;
+		public short extCode;
 	}
 }
